Add local storage script
Adds a script to be run in the head of each page to ensure local storage
data structure version compliance site wide. The script checks if ls
exists, if not it saves the current template data to ls, if it does the
script checks if updates are needed and applies them then saves back to
ls.